Add files via upload
This was a Result Analysis Project made to check or analyze the results of six divisions in the FE department. It has two options i.e. results of a complete division out of six and the result of a particular student. It also shows a graphical representation of pass, fail students in a division.